S3 read/write refactoring. (#328)
* Add key_suffix and key_ignore_suffix.

* Create _read_parallel.py.

* Parallelism applied to text readers.

* Testing refactoring done.
