+++ awswrangler/s3/_write_text.py
@@ -1,0 +1,516 @@
+"""Amazon CSV S3 Write Module (PRIVATE)."""
+
+import csv
+import logging
+import uuid
+from typing import Dict, List, Optional, Tuple, Union
+
+import boto3  # type: ignore
+import pandas as pd  # type: ignore
+import s3fs  # type: ignore
+
+from awswrangler import _data_types, _utils, catalog, exceptions
+from awswrangler.s3._delete import delete_objects
+from awswrangler.s3._write_concurrent import _write_proxy, _WriteProxy
+
+_logger: logging.Logger = logging.getLogger(__name__)
+
+
+def _to_text(
+    file_format: str,
+    df: pd.DataFrame,
+    path: str,
+    boto3_session: Optional[boto3.Session],
+    s3_additional_kwargs: Optional[Dict[str, str]],
+    **pandas_kwargs,
+) -> str:
+    if df.empty is True:
+        raise exceptions.EmptyDataFrame()
+    fs: s3fs.S3FileSystem = _utils.get_fs(
+        block_size=33_554_432, session=boto3_session, s3_additional_kwargs=s3_additional_kwargs  # 32 MB (32 * 2**20)
+    )
+    encoding: Optional[str] = pandas_kwargs.get("encoding", None)
+    newline: Optional[str] = pandas_kwargs.get("line_terminator", None)
+    with _utils.open_file(fs=fs, path=path, mode="w", encoding=encoding, newline=newline) as f:
+        if file_format == "csv":
+            df.to_csv(f, **pandas_kwargs)
+        elif file_format == "json":
+            df.to_json(f, **pandas_kwargs)
+    return path
+
+
+def _to_csv_dataset(
+    df: pd.DataFrame,
+    path: str,
+    index: bool,
+    sep: str,
+    use_threads: bool,
+    mode: str,
+    dtype: Dict[str, str],
+    partition_cols: Optional[List[str]],
+    boto3_session: Optional[boto3.Session],
+    s3_additional_kwargs: Optional[Dict[str, str]],
+) -> Tuple[List[str], Dict[str, List[str]]]:
+    paths: List[str] = []
+    partitions_values: Dict[str, List[str]] = {}
+    path = path if path[-1] == "/" else f"{path}/"
+    if mode not in ["append", "overwrite", "overwrite_partitions"]:
+        raise exceptions.InvalidArgumentValue(
+            f"{mode} is a invalid mode, please use append, overwrite or overwrite_partitions."
+        )
+    if (mode == "overwrite") or ((mode == "overwrite_partitions") and (not partition_cols)):
+        delete_objects(path=path, use_threads=use_threads, boto3_session=boto3_session)
+    df = _data_types.cast_pandas_with_athena_types(df=df, dtype=dtype)
+    _logger.debug("dtypes: %s", df.dtypes)
+    if not partition_cols:
+        file_path: str = f"{path}{uuid.uuid4().hex}.csv"
+        _to_text(
+            file_format="csv",
+            df=df,
+            path=file_path,
+            quoting=csv.QUOTE_NONE,
+            escapechar="\\",
+            header=False,
+            date_format="%Y-%m-%d %H:%M:%S.%f",
+            index=index,
+            sep=sep,
+            boto3_session=boto3_session,
+            s3_additional_kwargs=s3_additional_kwargs,
+        )
+        paths.append(file_path)
+    else:
+        proxy: _WriteProxy
+        with _write_proxy(use_threads=use_threads) as proxy:
+            for keys, subgroup in df.groupby(by=partition_cols, observed=True):
+                subgroup = subgroup.drop(partition_cols, axis="columns")
+                keys = (keys,) if not isinstance(keys, tuple) else keys
+                subdir = "/".join([f"{name}={val}" for name, val in zip(partition_cols, keys)])
+                prefix: str = f"{path}{subdir}/"
+                if mode == "overwrite_partitions":
+                    delete_objects(path=prefix, use_threads=use_threads, boto3_session=boto3_session)
+                file_path = f"{prefix}{uuid.uuid4().hex}.csv"
+                proxy.write(
+                    func=_to_text,
+                    boto3_session=boto3_session,
+                    file_format="csv",
+                    df=subgroup,
+                    path=file_path,
+                    quoting=csv.QUOTE_NONE,
+                    escapechar="\\",
+                    header=False,
+                    date_format="%Y-%m-%d %H:%M:%S.%f",
+                    index=index,
+                    sep=sep,
+                    s3_additional_kwargs=s3_additional_kwargs,
+                )
+                paths.append(file_path)
+                partitions_values[prefix] = [str(k) for k in keys]
+    return paths, partitions_values
+
+
+def to_csv(  # pylint: disable=too-many-arguments,too-many-locals
+    df: pd.DataFrame,
+    path: str,
+    sep: str = ",",
+    index: bool = True,
+    columns: Optional[List[str]] = None,
+    use_threads: bool = True,
+    boto3_session: Optional[boto3.Session] = None,
+    s3_additional_kwargs: Optional[Dict[str, str]] = None,
+    sanitize_columns: bool = False,
+    dataset: bool = False,
+    partition_cols: Optional[List[str]] = None,
+    mode: Optional[str] = None,
+    catalog_versioning: bool = False,
+    database: Optional[str] = None,
+    table: Optional[str] = None,
+    dtype: Optional[Dict[str, str]] = None,
+    description: Optional[str] = None,
+    parameters: Optional[Dict[str, str]] = None,
+    columns_comments: Optional[Dict[str, str]] = None,
+    regular_partitions: bool = True,
+    projection_enabled: bool = False,
+    projection_types: Optional[Dict[str, str]] = None,
+    projection_ranges: Optional[Dict[str, str]] = None,
+    projection_values: Optional[Dict[str, str]] = None,
+    projection_intervals: Optional[Dict[str, str]] = None,
+    projection_digits: Optional[Dict[str, str]] = None,
+    **pandas_kwargs,
+) -> Dict[str, Union[List[str], Dict[str, List[str]]]]:
+    """Write CSV file or dataset on Amazon S3.
+
+    The concept of Dataset goes beyond the simple idea of files and enable more
+    complex features like partitioning, casting and catalog integration (Amazon Athena/AWS Glue Catalog).
+
+    Note
+    ----
+    If `dataset=True` The table name and all column names will be automatically sanitized using
+    `wr.catalog.sanitize_table_name` and `wr.catalog.sanitize_column_name`.
+    Please, pass `sanitize_columns=True` to force the same behaviour for `dataset=False`.
+
+    Note
+    ----
+    If `dataset=True`, `pandas_kwargs` will be ignored due
+    restrictive quoting, date_format, escapechar, encoding, etc required by Athena/Glue Catalog.
+
+    Note
+    ----
+    By now Pandas does not support in-memory CSV compression.
+    https://github.com/pandas-dev/pandas/issues/22555
+    So the `compression` will not be supported on Wrangler too.
+
+    Note
+    ----
+    On `append` mode, the `parameters` will be upsert on an existing table.
+
+    Note
+    ----
+    In case of `use_threads=True` the number of threads
+    that will be spawned will be gotten from os.cpu_count().
+
+    Parameters
+    ----------
+    df: pandas.DataFrame
+        Pandas DataFrame https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.html
+    path : str
+        Amazon S3 path (e.g. s3://bucket/filename.csv).
+    sep : str
+        String of length 1. Field delimiter for the output file.
+    index : bool
+        Write row names (index).
+    columns : List[str], optional
+        Columns to write.
+    use_threads : bool
+        True to enable concurrent requests, False to disable multiple threads.
+        If enabled os.cpu_count() will be used as the max number of threads.
+    boto3_session : boto3.Session(), optional
+        Boto3 Session. The default boto3 Session will be used if boto3_session receive None.
+    s3_additional_kwargs:
+        Forward to s3fs, useful for server side encryption
+        https://s3fs.readthedocs.io/en/latest/#serverside-encryption
+    sanitize_columns : bool
+        True to sanitize columns names or False to keep it as is.
+        True value is forced if `dataset=True`.
+    dataset : bool
+        If True store a parquet dataset instead of a single file.
+        If True, enable all follow arguments:
+        partition_cols, mode, database, table, description, parameters, columns_comments, .
+    partition_cols: List[str], optional
+        List of column names that will be used to create partitions. Only takes effect if dataset=True.
+    mode : str, optional
+        ``append`` (Default), ``overwrite``, ``overwrite_partitions``. Only takes effect if dataset=True.
+    catalog_versioning : bool
+        If True and `mode="overwrite"`, creates an archived version of the table catalog before updating it.
+    database : str, optional
+        Glue/Athena catalog: Database name.
+    table : str, optional
+        Glue/Athena catalog: Table name.
+    dtype : Dict[str, str], optional
+        Dictionary of columns names and Athena/Glue types to be casted.
+        Useful when you have columns with undetermined or mixed data types.
+        (e.g. {'col name': 'bigint', 'col2 name': 'int'})
+    description : str, optional
+        Glue/Athena catalog: Table description
+    parameters : Dict[str, str], optional
+        Glue/Athena catalog: Key/value pairs to tag the table.
+    columns_comments : Dict[str, str], optional
+        Glue/Athena catalog:
+        Columns names and the related comments (e.g. {'col0': 'Column 0.', 'col1': 'Column 1.', 'col2': 'Partition.'}).
+    regular_partitions : bool
+        Create regular partitions (Non projected partitions) on Glue Catalog.
+        Disable when you will work only with Partition Projection.
+        Keep enabled even when working with projections is useful to keep
+        Redshift Spectrum working with the regular partitions.
+    projection_enabled : bool
+        Enable Partition Projection on Athena (https://docs.aws.amazon.com/athena/latest/ug/partition-projection.html)
+    projection_types : Optional[Dict[str, str]]
+        Dictionary of partitions names and Athena projections types.
+        Valid types: "enum", "integer", "date", "injected"
+        https://docs.aws.amazon.com/athena/latest/ug/partition-projection-supported-types.html
+        (e.g. {'col_name': 'enum', 'col2_name': 'integer'})
+    projection_ranges: Optional[Dict[str, str]]
+        Dictionary of partitions names and Athena projections ranges.
+        https://docs.aws.amazon.com/athena/latest/ug/partition-projection-supported-types.html
+        (e.g. {'col_name': '0,10', 'col2_name': '-1,8675309'})
+    projection_values: Optional[Dict[str, str]]
+        Dictionary of partitions names and Athena projections values.
+        https://docs.aws.amazon.com/athena/latest/ug/partition-projection-supported-types.html
+        (e.g. {'col_name': 'A,B,Unknown', 'col2_name': 'foo,boo,bar'})
+    projection_intervals: Optional[Dict[str, str]]
+        Dictionary of partitions names and Athena projections intervals.
+        https://docs.aws.amazon.com/athena/latest/ug/partition-projection-supported-types.html
+        (e.g. {'col_name': '1', 'col2_name': '5'})
+    projection_digits: Optional[Dict[str, str]]
+        Dictionary of partitions names and Athena projections digits.
+        https://docs.aws.amazon.com/athena/latest/ug/partition-projection-supported-types.html
+        (e.g. {'col_name': '1', 'col2_name': '2'})
+    pandas_kwargs :
+        keyword arguments forwarded to pandas.DataFrame.to_csv()
+        https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.to_csv.html
+
+    Returns
+    -------
+    None
+        None.
+
+    Examples
+    --------
+    Writing single file
+
+    >>> import awswrangler as wr
+    >>> import pandas as pd
+    >>> wr.s3.to_csv(
+    ...     df=pd.DataFrame({'col': [1, 2, 3]}),
+    ...     path='s3://bucket/prefix/my_file.csv',
+    ... )
+    {
+        'paths': ['s3://bucket/prefix/my_file.csv'],
+        'partitions_values': {}
+    }
+
+    Writing single file encrypted with a KMS key
+
+    >>> import awswrangler as wr
+    >>> import pandas as pd
+    >>> wr.s3.to_csv(
+    ...     df=pd.DataFrame({'col': [1, 2, 3]}),
+    ...     path='s3://bucket/prefix/my_file.csv',
+    ...     s3_additional_kwargs={
+    ...         'ServerSideEncryption': 'aws:kms',
+    ...         'SSEKMSKeyId': 'YOUR_KMY_KEY_ARN'
+    ...     }
+    ... )
+    {
+        'paths': ['s3://bucket/prefix/my_file.csv'],
+        'partitions_values': {}
+    }
+
+    Writing partitioned dataset
+
+    >>> import awswrangler as wr
+    >>> import pandas as pd
+    >>> wr.s3.to_csv(
+    ...     df=pd.DataFrame({
+    ...         'col': [1, 2, 3],
+    ...         'col2': ['A', 'A', 'B']
+    ...     }),
+    ...     path='s3://bucket/prefix',
+    ...     dataset=True,
+    ...     partition_cols=['col2']
+    ... )
+    {
+        'paths': ['s3://.../col2=A/x.csv', 's3://.../col2=B/y.csv'],
+        'partitions_values: {
+            's3://.../col2=A/': ['A'],
+            's3://.../col2=B/': ['B']
+        }
+    }
+
+    Writing dataset to S3 with metadata on Athena/Glue Catalog.
+
+    >>> import awswrangler as wr
+    >>> import pandas as pd
+    >>> wr.s3.to_csv(
+    ...     df=pd.DataFrame({
+    ...         'col': [1, 2, 3],
+    ...         'col2': ['A', 'A', 'B']
+    ...     }),
+    ...     path='s3://bucket/prefix',
+    ...     dataset=True,
+    ...     partition_cols=['col2'],
+    ...     database='default',  # Athena/Glue database
+    ...     table='my_table'  # Athena/Glue table
+    ... )
+    {
+        'paths': ['s3://.../col2=A/x.csv', 's3://.../col2=B/y.csv'],
+        'partitions_values: {
+            's3://.../col2=A/': ['A'],
+            's3://.../col2=B/': ['B']
+        }
+    }
+
+    Writing dataset casting empty column data type
+
+    >>> import awswrangler as wr
+    >>> import pandas as pd
+    >>> wr.s3.to_csv(
+    ...     df=pd.DataFrame({
+    ...         'col': [1, 2, 3],
+    ...         'col2': ['A', 'A', 'B'],
+    ...         'col3': [None, None, None]
+    ...     }),
+    ...     path='s3://bucket/prefix',
+    ...     dataset=True,
+    ...     database='default',  # Athena/Glue database
+    ...     table='my_table'  # Athena/Glue table
+    ...     dtype={'col3': 'date'}
+    ... )
+    {
+        'paths': ['s3://.../x.csv'],
+        'partitions_values: {}
+    }
+
+    """
+    if (database is None) ^ (table is None):
+        raise exceptions.InvalidArgumentCombination(
+            "Please pass database and table arguments to be able to store the metadata into the Athena/Glue Catalog."
+        )
+    if df.empty is True:
+        raise exceptions.EmptyDataFrame()
+
+    partition_cols = partition_cols if partition_cols else []
+    dtype = dtype if dtype else {}
+    partitions_values: Dict[str, List[str]] = {}
+
+    # Sanitize table to respect Athena's standards
+    if (sanitize_columns is True) or (dataset is True):
+        df = catalog.sanitize_dataframe_columns_names(df=df)
+        partition_cols = [catalog.sanitize_column_name(p) for p in partition_cols]
+        dtype = {catalog.sanitize_column_name(k): v.lower() for k, v in dtype.items()}
+        _utils.check_duplicated_columns(df=df)
+
+    session: boto3.Session = _utils.ensure_session(session=boto3_session)
+    if dataset is False:
+        if partition_cols:
+            raise exceptions.InvalidArgumentCombination("Please, pass dataset=True to be able to use partition_cols.")
+        if mode is not None:
+            raise exceptions.InvalidArgumentCombination("Please pass dataset=True to be able to use mode.")
+        if columns_comments:
+            raise exceptions.InvalidArgumentCombination("Please pass dataset=True to be able to use columns_comments.")
+        if any(arg is not None for arg in (database, table, description, parameters)):
+            raise exceptions.InvalidArgumentCombination(
+                "Please pass dataset=True to be able to use any one of these "
+                "arguments: database, table, description, parameters, "
+                "columns_comments."
+            )
+        pandas_kwargs["sep"] = sep
+        pandas_kwargs["index"] = index
+        pandas_kwargs["columns"] = columns
+        _to_text(
+            file_format="csv",
+            df=df,
+            path=path,
+            boto3_session=session,
+            s3_additional_kwargs=s3_additional_kwargs,
+            **pandas_kwargs,
+        )
+        paths = [path]
+    else:
+        mode = "append" if mode is None else mode
+        if columns:
+            df = df[columns]
+        if (
+            (mode in ("append", "overwrite_partitions")) and (database is not None) and (table is not None)
+        ):  # Fetching Catalog Types
+            catalog_types: Optional[Dict[str, str]] = catalog.get_table_types(
+                database=database, table=table, boto3_session=session
+            )
+            if catalog_types is not None:
+                for k, v in catalog_types.items():
+                    dtype[k] = v
+        paths, partitions_values = _to_csv_dataset(
+            df=df,
+            path=path,
+            index=index,
+            sep=sep,
+            use_threads=use_threads,
+            partition_cols=partition_cols,
+            dtype=dtype,
+            mode=mode,
+            boto3_session=session,
+            s3_additional_kwargs=s3_additional_kwargs,
+        )
+        if (database is not None) and (table is not None):
+            columns_types, partitions_types = _data_types.athena_types_from_pandas_partitioned(
+                df=df, index=index, partition_cols=partition_cols, dtype=dtype, index_left=True
+            )
+            catalog.create_csv_table(
+                database=database,
+                table=table,
+                path=path,
+                columns_types=columns_types,
+                partitions_types=partitions_types,
+                description=description,
+                parameters=parameters,
+                columns_comments=columns_comments,
+                boto3_session=session,
+                mode=mode,
+                catalog_versioning=catalog_versioning,
+                sep=sep,
+                projection_enabled=projection_enabled,
+                projection_types=projection_types,
+                projection_ranges=projection_ranges,
+                projection_values=projection_values,
+                projection_intervals=projection_intervals,
+                projection_digits=projection_digits,
+            )
+            if partitions_values and (regular_partitions is True):
+                _logger.debug("partitions_values:\n%s", partitions_values)
+                catalog.add_csv_partitions(
+                    database=database, table=table, partitions_values=partitions_values, boto3_session=session, sep=sep
+                )
+    return {"paths": paths, "partitions_values": partitions_values}
+
+
+def to_json(
+    df: pd.DataFrame,
+    path: str,
+    boto3_session: Optional[boto3.Session] = None,
+    s3_additional_kwargs: Optional[Dict[str, str]] = None,
+    **pandas_kwargs,
+) -> None:
+    """Write JSON file on Amazon S3.
+
+    Parameters
+    ----------
+    df: pandas.DataFrame
+        Pandas DataFrame https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.html
+    path : str
+        Amazon S3 path (e.g. s3://bucket/filename.csv).
+    boto3_session : boto3.Session(), optional
+        Boto3 Session. The default boto3 Session will be used if boto3_session receive None.
+    s3_additional_kwargs:
+        Forward to s3fs, useful for server side encryption
+        https://s3fs.readthedocs.io/en/latest/#serverside-encryption
+    pandas_kwargs:
+        keyword arguments forwarded to pandas.DataFrame.to_csv()
+        https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.to_json.html
+
+    Returns
+    -------
+    None
+        None.
+
+    Examples
+    --------
+    Writing JSON file
+
+    >>> import awswrangler as wr
+    >>> import pandas as pd
+    >>> wr.s3.to_json(
+    ...     df=pd.DataFrame({'col': [1, 2, 3]}),
+    ...     path='s3://bucket/filename.json',
+    ... )
+
+    Writing CSV file encrypted with a KMS key
+
+    >>> import awswrangler as wr
+    >>> import pandas as pd
+    >>> wr.s3.to_json(
+    ...     df=pd.DataFrame({'col': [1, 2, 3]}),
+    ...     path='s3://bucket/filename.json',
+    ...     s3_additional_kwargs={
+    ...         'ServerSideEncryption': 'aws:kms',
+    ...         'SSEKMSKeyId': 'YOUR_KMY_KEY_ARN'
+    ...     }
+    ... )
+
+    """
+    _to_text(
+        file_format="json",
+        df=df,
+        path=path,
+        boto3_session=boto3_session,
+        s3_additional_kwargs=s3_additional_kwargs,
+        **pandas_kwargs,
+    )
