+++ requirements-dev.txt
@@ -11,7 +11,6 @@ tox==3.17.1
 pytest==5.4.3
 pytest-cov==2.10.0
 pytest-xdist==1.33.0
-pytest-timeout==1.4.2
 scikit-learn==0.23.1
 cfn-lint==0.34.0
 cfn-flip==1.2.3
