+++ requirements.txt
@@ -1,10 +1,10 @@
 boto3>=1.12.0,<2.0.0
 botocore>=1.15.0,<2.0.0
 numpy>=1.18.0,<1.20.0
+pymysql>=0.9.0,<0.11.0
 sqlalchemy-redshift>=0.7.0,<0.9.0
 SQLAlchemy>=1.3.10,<1.3.14
 pandas~=1.0.0
 pyarrow~=0.17.0
 s3fs~=0.4.2
 psycopg2-binary~=2.8.0
-pymysql~=0.9.0
