+++ tox.ini
@@ -6,10 +6,9 @@ passenv = AWS_PROFILE AWS_DEFAULT_REGION AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY
 deps =
        pytest
        pytest-xdist
-       pytest-timeout
        moto
 commands =
-       pytest --timeout=900 -n 8 tests
+       pytest -n 16 tests
 
 [testenv:py36]
 passenv = AWS_PROFILE AWS_DEFAULT_REGION AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY
@@ -17,4 +16,4 @@ deps =
        {[testenv]deps}
        pytest-cov
 commands =
-       pytest --timeout=900 --cov=awswrangler -n 8 tests
+       pytest -n 16 --cov=awswrangler tests
